# apps/server/generated/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/sessions": {
      "post": {
        "operationId": "SessionController_start",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session started",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/StartSessionResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start a new study session",
        "tags": [
          "Learning"
        ]
      },
      "get": {
        "description": "Paginated. Optionally filter by studySetId to power the set page's activity list. Ordered newest-first by startedAt.",
        "operationId": "SessionController_list",
        "parameters": [
          {
            "name": "studySetId",
            "required": false,
            "in": "query",
            "description": "Optional — restrict history to sessions on this study set. Omit to return all of the user's sessions across sets.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sessions retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SessionHistoryItemDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List my study session history",
        "tags": [
          "Learning"
        ]
      }
    },
    "/sessions/{id}/answer": {
      "post": {
        "description": "Applies the outcome to the card mastery engine and returns the updated card + set progress.",
        "operationId": "SessionController_answer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAnswerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Answer recorded",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AnswerResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit an answer within a session",
        "tags": [
          "Learning"
        ]
      }
    },
    "/sessions/{id}/answer-written": {
      "post": {
        "description": "For Write / Learn-written / Test-written / AI-generated modes. The server runs the Levenshtein-tolerant evaluator against the card's canonical definition, translates the result into a mastery-engine outcome, and returns both the applied progress and the evaluation details (matchType, similarity, edit distance, normalized strings) so the frontend can render a diff and helpful feedback.",
        "operationId": "SessionController_answerWritten",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitWrittenAnswerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Written answer evaluated",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/WrittenAnswerResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit a typed answer — server evaluates and applies to mastery",
        "tags": [
          "Learning"
        ]
      }
    },
    "/sessions/{id}/next-batch": {
      "get": {
        "description": "Returns 7–10 non-mastered cards mixed between multiple-choice (recognition) and free-text (recall) prompts based on each card's current weighted streak. Only supported for sessions started with mode LEARN.",
        "operationId": "SessionController_nextBatch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "description": "Target number of cards to include in the batch.",
            "schema": {
              "minimum": 1,
              "maximum": 20,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Batch retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/LearnBatchResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch the next batch of Learn Mode cards",
        "tags": [
          "Learning"
        ]
      }
    },
    "/sessions/{id}/complete": {
      "post": {
        "operationId": "SessionController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session completed",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SessionSummaryDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a session complete and return the summary",
        "tags": [
          "Learning"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mimir API",
    "description": "Auto-generated OpenAPI specification for the Mimir backend. This file is the source of truth for the frontend type pipeline.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt"
      }
    },
    "schemas": {
...
      "ResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "data": {
            "type": "object"
          },
          "error": {
            "type": "string"
          },
          "meta": {
            "type": "object"
          }
        },
        "required": [
          "ok",
          "message"
        ]
      },
...
      "StartSessionResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "studySetId": {
            "type": "string",
            "format": "uuid"
          },
          "mode": {
            "type": "string",
            "enum": [
              "FLASHCARD",
              "LEARN",
              "WRITE",
              "SPELL",
              "TEST",
              "MATCH",
              "AI_FILL_BLANK",
              "AI_GUESS_WORD"
            ]
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "sessionId",
          "studySetId",
          "mode",
          "startedAt"
        ]
      },
      "StartSessionDto": {
        "type": "object",
        "properties": {
          "studySetId": {
            "type": "string",
            "format": "uuid"
          },
          "mode": {
            "type": "string",
            "enum": [
              "FLASHCARD",
              "LEARN",
              "WRITE",
              "SPELL",
              "TEST",
              "MATCH",
              "AI_FILL_BLANK",
              "AI_GUESS_WORD"
            ]
          }
        },
        "required": [
          "studySetId",
          "mode"
        ]
      },
      "SessionHistoryItemDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "studySetId": {
            "type": "string",
            "format": "uuid"
          },
          "mode": {
            "type": "string",
            "enum": [
              "FLASHCARD",
              "LEARN",
              "WRITE",
              "SPELL",
              "TEST",
              "MATCH",
              "AI_FILL_BLANK",
              "AI_GUESS_WORD"
            ]
          },
          "cardsStudied": {
            "type": "number"
          },
          "correctAnswers": {
            "type": "number"
          },
          "incorrectAnswers": {
            "type": "number"
          },
          "durationSeconds": {
            "type": "number"
          },
          "accuracy": {
            "type": "number",
            "description": "Accuracy as a fraction (0.00–1.00). Zero if the session recorded no correct/incorrect answers yet.",
            "example": 0.85
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Null while a session is still in progress."
          }
        },
        "required": [
          "sessionId",
          "studySetId",
          "mode",
          "cardsStudied",
          "correctAnswers",
          "incorrectAnswers",
          "durationSeconds",
          "accuracy",
          "startedAt"
        ]
      },
      "CardProgressSummaryDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "LEARNING",
              "MASTERED"
            ]
          },
          "weightedStreak": {
            "type": "string",
            "example": "1.50"
          },
          "correctCount": {
            "type": "number"
          },
          "incorrectCount": {
            "type": "number"
          },
          "masteredAt": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "status",
          "weightedStreak",
          "correctCount",
          "incorrectCount"
        ]
      },
      "SetProgressSummaryDto": {
        "type": "object",
        "properties": {
          "totalCards": {
            "type": "number"
          },
          "newCount": {
            "type": "number"
          },
          "learningCount": {
            "type": "number"
          },
          "masteredCount": {
            "type": "number"
          }
        },
        "required": [
          "totalCards",
          "newCount",
          "learningCount",
          "masteredCount"
        ]
      },
      "AnswerResponseDto": {
        "type": "object",
        "properties": {
          "correct": {
            "type": "boolean",
            "description": "True if the submitted outcome was CORRECT (echoed for UI convenience)."
          },
          "correctAnswer": {
            "type": "string",
            "description": "Canonical answer text for the card."
          },
          "graduated": {
            "type": "boolean",
            "description": "Whether this attempt caused the card to reach MASTERED for the first time. SRS row created if true."
          },
          "demoted": {
            "type": "boolean",
            "description": "Whether this attempt demoted a previously MASTERED card back to LEARNING."
          },
          "cardProgress": {
            "$ref": "#/components/schemas/CardProgressSummaryDto"
          },
          "setProgress": {
            "$ref": "#/components/schemas/SetProgressSummaryDto"
          }
        },
        "required": [
          "correct",
          "correctAnswer",
          "graduated",
          "demoted",
          "cardProgress",
          "setProgress"
        ]
      },
      "SubmitAnswerDto": {
        "type": "object",
        "properties": {
          "attemptId": {
            "type": "string",
            "format": "uuid",
            "description": "Client-generated UUID used to make retries idempotent. Second submit with the same attemptId returns the first result without re-applying it."
          },
          "cardId": {
            "type": "string",
            "format": "uuid"
          },
          "studyMode": {
            "type": "string",
            "enum": [
              "FLASHCARD",
              "LEARN_MC",
              "LEARN_WRITTEN",
              "WRITE",
              "SPELL",
              "TEST_WRITTEN",
              "TEST_MC",
              "TEST_TF",
              "AI_FILL_BLANK",
              "AI_GUESS_WORD",
              "MATCH"
            ]
          },
          "outcome": {
            "type": "string",
            "enum": [
              "CORRECT",
              "INCORRECT",
              "SKIPPED"
            ]
          },
          "hintUsed": {
            "type": "boolean",
            "example": false
          },
          "response": {
            "type": "string",
            "description": "The learner-typed response for written modes. Used to echo back on the client; not stored on the mastery row.",
            "maxLength": 500
          }
        },
        "required": [
          "attemptId",
          "cardId",
          "studyMode",
          "outcome",
          "hintUsed"
        ]
      },
      "WriteEvaluationDto": {
        "type": "object",
        "properties": {
          "matchType": {
            "type": "string",
            "enum": [
              "EXACT",
              "TYPO_ACCEPTED",
              "WRONG"
            ],
            "description": "EXACT = normalized strings match. TYPO_ACCEPTED = Levenshtein distance ≤ 1 in a 6+ character answer (still scored CORRECT). WRONG = anything else."
          },
          "similarity": {
            "type": "number",
            "description": "0.00–1.00. Computed as 1 - editDistance / max(len). Frontend can render a \"you were close\" screen using this even when the outcome is INCORRECT.",
            "example": 0.92
          },
          "editDistance": {
            "type": "number",
            "description": "Levenshtein distance between the normalized strings."
          },
          "normalizedInput": {
            "type": "string",
            "description": "The learner's answer after normalization. Useful for client-side diff rendering."
          },
          "normalizedExpected": {
            "type": "string",
            "description": "The canonical answer after normalization."
          }
        },
        "required": [
          "matchType",
          "similarity",
          "editDistance",
          "normalizedInput",
          "normalizedExpected"
        ]
      },
      "WrittenAnswerResponseDto": {
        "type": "object",
        "properties": {
          "correct": {
            "type": "boolean"
          },
          "correctAnswer": {
            "type": "string",
            "description": "Canonical answer text for the card."
          },
          "graduated": {
            "type": "boolean"
          },
          "demoted": {
            "type": "boolean"
          },
          "cardProgress": {
            "$ref": "#/components/schemas/CardProgressSummaryDto"
          },
          "setProgress": {
            "$ref": "#/components/schemas/SetProgressSummaryDto"
          },
          "evaluation": {
            "$ref": "#/components/schemas/WriteEvaluationDto"
          }
        },
        "required": [
          "correct",
          "correctAnswer",
          "graduated",
          "demoted",
          "cardProgress",
          "setProgress",
          "evaluation"
        ]
      },
      "SubmitWrittenAnswerDto": {
        "type": "object",
        "properties": {
          "attemptId": {
            "type": "string",
            "format": "uuid",
            "description": "Client-generated UUID used to make retries idempotent. Second submit with the same attemptId returns the first result without re-applying it."
          },
          "cardId": {
            "type": "string",
            "format": "uuid"
          },
          "studyMode": {
            "type": "string",
            "enum": [
              "WRITE",
              "LEARN_WRITTEN",
              "TEST_WRITTEN",
              "AI_FILL_BLANK",
              "AI_GUESS_WORD"
            ],
            "description": "Which written mode produced this answer. Must be one of the recall-oriented modes; the multiple-choice / self-report modes go through POST /sessions/:id/answer directly."
          },
          "userAnswer": {
            "type": "string",
            "description": "The learner-typed answer, exactly as entered.",
            "maxLength": 500
          },
          "hintUsed": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "attemptId",
          "cardId",
          "studyMode",
          "userAnswer",
          "hintUsed"
        ]
      },
      "LearnBatchCardDto": {
        "type": "object",
        "properties": {
          "cardId": {
            "type": "string",
            "format": "uuid"
          },
          "term": {
            "type": "string",
            "description": "The term shown to the learner as the prompt."
          },
          "hint": {
            "type": "string",
            "nullable": true,
            "description": "Optional hint attached to the card. Frontend only shows this if the learner clicks the hint button."
          },
          "promptType": {
            "type": "string",
            "enum": [
              "LEARN_MC",
              "LEARN_WRITTEN"
            ],
            "description": "How the frontend should prompt: multiple choice (recognition) or free-text (recall). Chosen server-side based on the learner's current mastery streak."
          },
          "choices": {
            "description": "Four shuffled options for LEARN_MC prompts. Undefined for LEARN_WRITTEN.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "correctChoiceIndex": {
            "type": "number",
            "description": "Index of the correct choice for LEARN_MC prompts. Same trust model as Flashcard Mode — the client uses this to render feedback."
          }
        },
        "required": [
          "cardId",
          "term",
          "promptType"
        ]
      },
      "LearnBatchResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "cards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LearnBatchCardDto"
            }
          },
          "hasMoreCards": {
            "type": "boolean",
            "description": "True if more non-mastered cards remain in the set beyond this batch. When false, all cards have reached MASTERED and the frontend can offer a completion screen."
          }
        },
        "required": [
          "sessionId",
          "cards",
          "hasMoreCards"
        ]
      },
      "SessionSummaryDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "studySetId": {
            "type": "string",
            "format": "uuid"
          },
          "mode": {
            "type": "string",
            "enum": [
              "FLASHCARD",
              "LEARN",
              "WRITE",
              "SPELL",
              "TEST",
              "MATCH",
              "AI_FILL_BLANK",
              "AI_GUESS_WORD"
            ]
          },
          "cardsStudied": {
            "type": "number"
          },
          "correctAnswers": {
            "type": "number"
          },
          "incorrectAnswers": {
            "type": "number"
          },
          "durationSeconds": {
            "type": "number",
            "description": "Total wall-clock time in seconds from startedAt to completedAt."
          },
          "accuracy": {
            "type": "number",
            "description": "Accuracy as a fraction (0.00–1.00). Zero if no correct+incorrect answers recorded.",
            "example": 0.85
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "sessionId",
          "studySetId",
          "mode",
          "cardsStudied",
          "correctAnswers",
          "incorrectAnswers",
          "durationSeconds",
          "accuracy",
          "startedAt",
          "completedAt"
        ]
      }
    }
  }
}